move wallet resolver to ens
